chore: migrate to MUSA platfrom

+++ transformer_engine/common/util/padding.cu
@@ -4,7 +4,9 @@
  * See LICENSE for license information.
  ************************************************************************/
 
-#include <cuda_runtime.h>
+ #include "transformer_engine/musify.h"
+ 
+#include <musa_runtime.h>
 #include <transformer_engine/padding.h>
 
 #include <cfloat>
@@ -101,7 +103,7 @@ __global__ void __launch_bounds__(threads_per_block) multi_padding_kernel(MultiP
       if (row < num_rows) {
         for (int j2 = 0; j2 < nvec; ++j2) {
           if (col + j2 < row_length) {
-            local_input.data.elt[j2] = input[static_cast<size_t>(row) * row_length + col + j2];
+            local_input.data.elt[j2] = input[row * row_length + col + j2];
           }
         }
       }
@@ -112,14 +114,14 @@ __global__ void __launch_bounds__(threads_per_block) multi_padding_kernel(MultiP
       if (row < num_rows) {
         for (int j2 = 0; j2 < nvec; ++j2) {
           if (col + j2 < row_length) {
-            output[static_cast<size_t>(row) * row_length + col + j2] = local_output.data.elt[j2];
+            output[row * row_length + col + j2] = local_output.data.elt[j2];
           }
         }
       } else if (row < padded_num_rows) {
         // padding
         for (int j2 = 0; j2 < nvec; ++j2) {
           if (col + j2 < row_length) {
-            output[static_cast<size_t>(row) * row_length + col + j2] = local_zero;
+            output[row * row_length + col + j2] = local_zero;
           }
         }
       }
@@ -185,7 +187,7 @@ __global__ void __launch_bounds__(threads_per_block) multi_unpadding_kernel(Mult
       if (row < num_rows) {
         for (int j2 = 0; j2 < nvec; ++j2) {
           if (col + j2 < row_length) {
-            local_input.data.elt[j2] = input[static_cast<size_t>(row) * row_length + col + j2];
+            local_input.data.elt[j2] = input[row * row_length + col + j2];
           }
         }
       }
@@ -196,7 +198,7 @@ __global__ void __launch_bounds__(threads_per_block) multi_unpadding_kernel(Mult
       if (row < num_rows) {
         for (int j2 = 0; j2 < nvec; ++j2) {
           if (col + j2 < row_length) {
-            output[static_cast<size_t>(row) * row_length + col + j2] = local_output.data.elt[j2];
+            output[row * row_length + col + j2] = local_output.data.elt[j2];
           }
         }
       }
